Move AsCSV function from pkg/client/cli/util to pkg/slice
